# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Machine API
  version: 1.0.0
  description: API for managing coffee machine, drinks and orders

servers:
  - url: http://localhost:8080

tags:
  - name: Drinks
  - name: Orders
  - name: Machine

paths:
  /api/drinks:
    get:
      tags: [Drinks]
      summary: Get all drinks
      responses:
        '200':
          description: List of drinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drink'

  /api/drinks/{id}:
    get:
      tags: [Drinks]
      summary: Get drink by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Drink found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '404':
          $ref: '#/components/responses/Error'

  /api/drinks/popular:
    get:
      tags: [Drinks]
      summary: Get most popular drinks
      responses:
        '200':
          description: Popular drinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drink'

  /api/drinks/create:
    post:
      tags: [Drinks]
      summary: Create new drink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDrinkRequest'
      responses:
        '200':
          description: Drink created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'

  /api/drinks/update/{id}:
    put:
      tags: [Drinks]
      summary: Update drink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDrinkRequest'
      responses:
        '200':
          description: Updated drink
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'

  /api/machine/inventory:
    get:
      tags: [Machine]
      summary: Get machine inventory
      responses:
        '200':
          description: Inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'

  /api/machine/update:
    put:
      tags: [Machine]
      summary: Update machine inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryRequest'
      responses:
        '200':
          description: Updated inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'

  /api/order:
    get:
      tags: [Orders]
      summary: Order drink
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Drink prepared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '409':
          $ref: '#/components/responses/Error'

components:

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    Drink:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        waterAmount:
          type: integer
        coffeeAmount:
          type: integer
        milkAmount:
          type: integer

    CreateDrinkRequest:
      type: object
      required: [name, waterAmount, coffeeAmount, milkAmount]
      properties:
        name:
          type: string
        waterAmount:
          type: integer
        coffeeAmount:
          type: integer
        milkAmount:
          type: integer

    Inventory:
      type: object
      properties:
        water:
          type: integer
        coffee:
          type: integer
        milk:
          type: integer

    UpdateInventoryRequest:
      type: object
      properties:
        water:
          type: integer
        coffee:
          type: integer
        milk:
          type: integer

    ErrorResponse:
      type: object
      properties:
        message:
          type: string